+++ tests/Bridge/Symfony/DependencyInjection/Compiler/MapperCompilerPassTest.php
@@ -15,6 +15,11 @@ namespace Sonata\Doctrine\Tests\Bridge\Symfony\DependencyInjection\Compiler;
 
 use Matthias\SymfonyDependencyInjectionTest\PhpUnit\AbstractCompilerPassTestCase;
 use Sonata\Doctrine\Bridge\Symfony\DependencyInjection\Compiler\MapperCompilerPass;
+use Sonata\Doctrine\Mapper\Builder\OptionsBuilder;
+use Sonata\Doctrine\Mapper\DoctrineCollector;
+use Sonata\Doctrine\Mapper\ORM\DoctrineORMMapper;
+use Sonata\Doctrine\Tests\App\Entity\TestEntity;
+use Sonata\Doctrine\Tests\App\Entity\TestRelatedEntity;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
 
 /**
@@ -60,5 +65,33 @@ final class MapperCompilerPassTest extends AbstractCompilerPassTestCase
         $this->compile();
 
         $this->assertContainerBuilderHasService('sonata.doctrine.mapper');
+    }
+
+    public function testAssociationMapping(): void
+    {
+        $definition = $this->registerService('sonata.doctrine.mapper', DoctrineORMMapper::class);
+        $definition->setPublic(true);
+
+        $this->registerService('doctrine', 'foo');
+
+        $options = OptionsBuilder::createManyToOne('relation', TestRelatedEntity::class)
+            ->add('joinColumns', [['referencedColumnName' => 'id']]);
+
+        $collector = DoctrineCollector::getInstance();
+        $collector->addAssociation(TestEntity::class, 'mapManyToOne', $options);
+
+        $this->compile();
+
+        $compiledMapper = $this->container->get('sonata.doctrine.mapper');
+
+        $this->assertContainerBuilderHasServiceDefinitionWithMethodCall(
+            'sonata.doctrine.mapper',
+            'addAssociation',
+            [TestEntity::class, 'mapManyToOne', [$options->getOptions()]]
+        );
+
+        $this->assertInstanceOf(DoctrineORMMapper::class, $compiledMapper);
+
+        $collector->clear();
     }
 }